+++ changelog.md
@@ -7,6 +7,7 @@ This is what we are currently working on:
 0.9.9
 
 - Adding @neck/rinickolous's inventory sheet
+- Replaced /light dialog with parameters torch = /light 6 2 torch (or candle = /light 1 0 t)
 
 ## History